backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
asgiref==3.10.0
attrs==25.4.0
black==23.12.0
click==8.3.0
Django==4.2.7
django-cors-headers==4.3.1
django-mptt==0.16.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
factory-boy==3.3.0
Faker==37.11.0
flake8==6.1.0
frozenlist==1.8.0
idna==3.11
inflection==0.5.1
iniconfig==2.1.0
isort==5.13.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mccabe==0.7.0
multidict==6.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
yarl==1.22.0
